requirements.txt
Appium-Python-Client
pytest
selenium
time
datetime